+++ include/cloudbreak.bash
@@ -43,3 +43,107 @@ cloudbreak-conf-cbdb() {
 cloudbreak-conf-defaults() {
     env-import CB_BLUEPRINT_DEFAULTS "lambda-architecture,multi-node-hdfs-yarn,hdp-multinode-default"
 }
+
+check_if_running() {
+    declare desc="Checks if a container is in running state"
+
+    declare container="$1"
+
+    [[ "true" == $(docker inspect -f '{{.State.Running}}' $container 2>/dev/null) ]]
+}
+
+check_skip() {
+    declare desc="Checks if a container should be skipped: SKIP_XXX"
+    declare name=${1:?}
+    
+    if grep -qi $name <<< "${!SKIP_*}" && ! check_if_running $name; then
+        error  "$name should be skipped but its not running ?!"
+        exit 1
+    fi
+    
+    grep -qi $name <<< "${!SKIP_*}"
+}
+
+start_consul() {
+    declare desc="starts consul binding to: $PRIVATE_IP http:8500 dns:53 rpc:8400"
+
+    local name=consul
+    env-import PRIVATE_IP
+    
+    if check_skip $name; then
+        warn "skipping container: $name"
+        return
+    fi
+
+    if check_if_running $name; then
+        warn "$name is already running, not starting"
+        return
+    fi
+
+    info $desc
+    docker run -d \
+        -h node1 \
+        --name=$name \
+        --privileged \
+        -e SERVICE_IGNORE=true \
+        -p ${PRIVATE_IP}:53:53/udp \
+        -p ${PRIVATE_IP}:8400:8400 \
+        -p ${PRIVATE_IP}:8500:8500 \
+        -v /var/run/docker.sock:/var/run/docker.sock \
+        sequenceiq/consul:$DOCKER_TAG_CONSUL -server -bootstrap -advertise ${PRIVATE_IP}
+}
+
+start_registrator() {
+    declare desc="starts registrator connecting to consul"
+
+    local name=registrator
+    env-import PRIVATE_IP
+    
+    if check_skip $name; then
+        warn "skipping container: $name"
+        return
+    fi
+
+    if check_if_running $name; then
+        warn "$name is already running, not starting"
+        return
+    fi
+
+    info $desc
+    docker run -d \
+      --name=$name \
+      --privileged \
+      -v /var/run/docker.sock:/tmp/docker.sock \
+      gliderlabs/registrator:$DOCKER_TAG_REGISTRATOR consul://${PRIVATE_IP}:8500
+}
+
+cloudbreak-deploy() {
+    declare desc="Deploys all Cloudbreak components in docker containers"
+
+    start_consul | gray
+    start_registrator | gray
+}
+
+cloudbreak-destroy() {
+    declare desc="Destroys Cloudbreak related containers"
+
+    local containers="consul registrator"
+
+    info "killing containers: $containers"
+    for name in $containers; do
+        if ! docker inspect $name &>/dev/null ;then
+            warn "no such container: $name"
+        else 
+            if grep -qi $name <<< "${!SKIP_*}"; then
+                warn skipping: $name
+            else
+                debug stop: $name
+                docker stop -t 0 $name | gray
+                debug remove: $name
+                docker rm -f $name | gray
+                
+            fi
+        fi
+    done
+
+}