+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 - `make tests` runs unit tests
 - docker unit tests are added
+- start command added: WIP consul, registrator starts
+- kill command addd: stops and removes cloudbreak specific containers
+- SKIP_XXX skips the container start
 
 ### Removed
 
@@ -15,6 +18,7 @@
 - env export: machine friendly config list
 - env show: human readable config list
 - circle runs unit tests
+- snapshot binaries include branch name in version string
 
 ## [v0.0.6] - 2015-03-25